requirements.txt
configparser==3.7.3
Django==2.1
django-cors-headers==2.4.1
django-debug-toolbar==2.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
pre-commit==1.21.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pyparsing==2.4.6
pytest==4.6.3
pytest-cov==2.7.1
pytest-django==3.5.1
whitenoise==3.3.1
